Migrate navbar logic to seperate component of components folder
